feat: move request and response tests
