+++ crates/rpc/src/jsonrpc/request.rs
@@ -77,3 +77,114 @@ impl<'de> Deserialize<'de> for RpcRequest<'de> {
         })
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use serde_json::json;
+
+    use super::*;
+
+    #[test]
+    fn with_null_id() {
+        let json = json!({
+            "jsonrpc": "2.0",
+            "method": "sum",
+            "params": [1,2,3],
+            "id": null
+        });
+        let result = RpcRequest::deserialize(json).unwrap();
+        let expected = RpcRequest {
+            method: "sum".to_owned(),
+            params: json!([1, 2, 3]),
+            id: RequestId::Null,
+        };
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn with_string_id() {
+        let json = json!({
+            "jsonrpc": "2.0",
+            "method": "sum",
+            "params": [1,2,3],
+            "id": "text"
+        });
+        let result = RpcRequest::deserialize(json).unwrap();
+        let expected = RpcRequest {
+            method: "sum".to_owned(),
+            params: json!([1, 2, 3]),
+            id: RequestId::String("text".into()),
+        };
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn with_number_id() {
+        let json = json!({
+            "jsonrpc": "2.0",
+            "method": "sum",
+            "params": [1,2,3],
+            "id": 456
+        });
+        let result = RpcRequest::deserialize(json).unwrap();
+        let expected = RpcRequest {
+            method: "sum".to_owned(),
+            params: json!([1, 2, 3]),
+            id: RequestId::Number(456),
+        };
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn notification() {
+        let json = json!({
+            "jsonrpc": "2.0",
+            "method": "sum",
+            "params": [1,2,3]
+        });
+        let result = RpcRequest::deserialize(json).unwrap();
+        let expected = RpcRequest {
+            method: "sum".to_owned(),
+            params: json!([1, 2, 3]),
+            id: RequestId::Notification,
+        };
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn jsonrpc_version_missing() {
+        let json = json!({
+            "method": "sum",
+            "params": [1,2,3],
+            "id": 456
+        });
+        RpcRequest::deserialize(json).unwrap_err();
+    }
+
+    #[test]
+    fn jsonrpc_version_is_not_2() {
+        let json = json!({
+            "jsonrpc": "1.0",
+            "method": "sum",
+            "params": [1,2,3],
+            "id": 456
+        });
+        RpcRequest::deserialize(json).unwrap_err();
+    }
+
+    #[test]
+    fn no_params() {
+        let json = json!({
+            "jsonrpc": "2.0",
+            "method": "sum",
+            "id": 456
+        });
+        let result = RpcRequest::deserialize(json).unwrap();
+        let expected = RpcRequest {
+            method: "sum".to_owned(),
+            params: json!(null),
+            id: RequestId::Number(456),
+        };
+        assert_eq!(result, expected);
+    }
+}
