+++ crates/rpc/src/jsonrpc/response.rs
@@ -51,8 +51,7 @@ impl Serialize for RpcResponse<'_> {
             RequestId::Number(x) => obj.serialize_entry("id", &x)?,
             RequestId::String(x) => obj.serialize_entry("id", &x)?,
             RequestId::Null => obj.serialize_entry("id", &Value::Null)?,
-            RequestId::Notification => {},
-            
+            RequestId::Notification => {}
         };
 
         obj.end()
@@ -62,5 +61,49 @@ impl Serialize for RpcResponse<'_> {
 impl IntoResponse for RpcResponse<'_> {
     fn into_response(self) -> axum::response::Response {
         serde_json::to_vec(&self).unwrap().into_response()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use serde_json::json;
+
+    use super::*;
+
+    #[test]
+    fn output_is_error() {
+        let serialized = serde_json::to_value(&RpcResponse {
+            output: Err(RpcError::InvalidParams),
+            id: RequestId::Number(1),
+        })
+        .unwrap();
+
+        let expected = json!({
+            "jsonrpc": "2.0",
+            "error": {
+                "code": RpcError::InvalidParams.code(),
+                "message": RpcError::InvalidParams.message(),
+            },
+            "id": 1,
+        });
+
+        assert_eq!(serialized, expected);
+    }
+
+    #[test]
+    fn output_is_ok() {
+        let serialized = serde_json::to_value(&RpcResponse {
+            output: Ok(Value::String("foobar".to_owned())),
+            id: RequestId::Number(1),
+        })
+        .unwrap();
+
+        let expected = json!({
+            "jsonrpc": "2.0",
+            "result": "foobar",
+            "id": 1,
+        });
+
+        assert_eq!(serialized, expected);
     }
 }
