+++ crates/rpc/src/jsonrpc/router.rs
@@ -445,7 +445,6 @@ pub trait RpcMethodHandler {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::jsonrpc::RequestId;
     use serde::Deserialize;
     use serde_json::json;
 
@@ -804,155 +803,6 @@ mod tests {
                 .unwrap();
 
             assert_eq!(res.content_length(), Some(0));
-        }
-    }
-
-    mod request {
-        use super::*;
-
-        #[test]
-        fn with_null_id() {
-            let json = json!({
-                "jsonrpc": "2.0",
-                "method": "sum",
-                "params": [1,2,3],
-                "id": null
-            });
-            let result = RpcRequest::deserialize(json).unwrap();
-            let expected = RpcRequest {
-                method: "sum".to_owned(),
-                params: json!([1, 2, 3]),
-                id: RequestId::Null,
-            };
-            assert_eq!(result, expected);
-        }
-
-        #[test]
-        fn with_string_id() {
-            let json = json!({
-                "jsonrpc": "2.0",
-                "method": "sum",
-                "params": [1,2,3],
-                "id": "text"
-            });
-            let result = RpcRequest::deserialize(json).unwrap();
-            let expected = RpcRequest {
-                method: "sum".to_owned(),
-                params: json!([1, 2, 3]),
-                id: RequestId::String("text".into()),
-            };
-            assert_eq!(result, expected);
-        }
-
-        #[test]
-        fn with_number_id() {
-            let json = json!({
-                "jsonrpc": "2.0",
-                "method": "sum",
-                "params": [1,2,3],
-                "id": 456
-            });
-            let result = RpcRequest::deserialize(json).unwrap();
-            let expected = RpcRequest {
-                method: "sum".to_owned(),
-                params: json!([1, 2, 3]),
-                id: RequestId::Number(456),
-            };
-            assert_eq!(result, expected);
-        }
-
-        #[test]
-        fn notification() {
-            let json = json!({
-                "jsonrpc": "2.0",
-                "method": "sum",
-                "params": [1,2,3]
-            });
-            let result = RpcRequest::deserialize(json).unwrap();
-            let expected = RpcRequest {
-                method: "sum".to_owned(),
-                params: json!([1, 2, 3]),
-                id: RequestId::Notification,
-            };
-            assert_eq!(result, expected);
-        }
-
-        #[test]
-        fn jsonrpc_version_missing() {
-            let json = json!({
-                "method": "sum",
-                "params": [1,2,3],
-                "id": 456
-            });
-            RpcRequest::deserialize(json).unwrap_err();
-        }
-
-        #[test]
-        fn jsonrpc_version_is_not_2() {
-            let json = json!({
-                "jsonrpc": "1.0",
-                "method": "sum",
-                "params": [1,2,3],
-                "id": 456
-            });
-            RpcRequest::deserialize(json).unwrap_err();
-        }
-
-        #[test]
-        fn no_params() {
-            let json = json!({
-                "jsonrpc": "2.0",
-                "method": "sum",
-                "id": 456
-            });
-            let result = RpcRequest::deserialize(json).unwrap();
-            let expected = RpcRequest {
-                method: "sum".to_owned(),
-                params: json!(null),
-                id: RequestId::Number(456),
-            };
-            assert_eq!(result, expected);
-        }
-    }
-
-    mod response {
-        use super::*;
-
-        #[test]
-        fn output_is_error() {
-            let serialized = serde_json::to_value(&RpcResponse {
-                output: Err(RpcError::InvalidParams),
-                id: RequestId::Number(1),
-            })
-            .unwrap();
-
-            let expected = json!({
-                "jsonrpc": "2.0",
-                "error": {
-                    "code": RpcError::InvalidParams.code(),
-                    "message": RpcError::InvalidParams.message(),
-                },
-                "id": 1,
-            });
-
-            assert_eq!(serialized, expected);
-        }
-
-        #[test]
-        fn output_is_ok() {
-            let serialized = serde_json::to_value(&RpcResponse {
-                output: Ok(Value::String("foobar".to_owned())),
-                id: RequestId::Number(1),
-            })
-            .unwrap();
-
-            let expected = json!({
-                "jsonrpc": "2.0",
-                "result": "foobar",
-                "id": 1,
-            });
-
-            assert_eq!(serialized, expected);
         }
     }
